cairo: fix FT_Face reference counting
Both Cairo and FreeType have reference counting for their font
objects. On top of this, Cairo has a global cache for fonts, which
survives across contexts and surfaces. The API docs suggest to add a
finalizer to the Cairo face, but at the same time suggest that our
previous code was correct:

    You must not call FT_Done_Face() before the last reference to the
    cairo_font_face_t has been dropped.

I think we did this (~Cairo_outputter drops all cairo_font_face_ts), but
it seems the following was happening:

Cairo ingests a reference to the FT_Face without incrementing the
FT_Face reference counter *and* without setting up a finalizer for the
Cairo face.

As a result, the music font worked fine (We don't drop it unless we
clear the lilypond global font cache), but we dropped the FT_Faces we
loaded (ie. text fonts) in Cairo_outputter::~Cairo_outputter(), which
led to problems if cairo cached it across context/surface invocations.

I can't explain why this did not lead to OOB/UAF problems.

Fix this by:

1) attach finalizer to the Cairo face.  This will drop the FT_Face
count when the Cairo face dies.

2) call FT_Reference_Face: this ensures we don't kill the music font
if Cairo drops its reference to the font.

This seems to fix issue #6170, even though I'm not 100% sure of the
mechanics.