Markup commands for conditionals
This provides dedicated support for conditionals, which previously had
to use \on-the-fly with functions following the pattern:

   (lambda (layout props arg)
     (if (some-condition layout props)
         (interpret-markup layout props arg)
         empty-stencil))

The \if markup command is introduced.  Its first argument is a procedure
taking layout and props arguments, and returning a Scheme value, which
is interpreted as a boolean in Scheme's usual way.  The \unless command is
similar, but negates the condition.  The \if-else command is a ternary
operator.  Functions from titling-init.ly are refactored to take just
layout and props arguments and return booleans instead of interpreting
markups.  While at changing syntax, some of them are renamed for more
consistency.  The advantage is that the syntax is more intuitive (a
backslash in all cases, with \on-page being defined with
define-scheme-function) and it's easy to negate a condition (instead of
defining ad-hoc functions).

Closes #2576