Fix not-single-page
It was broken, not printing for the first page in the book even if there
was more than one page.  This seems to have been caused by the renaming
in commit 04c60c4727.  Just redefine it in terms of book-first-page? and
book-last-page?.  A test is added in the next commit, where it is
converted to simply return a boolean along with other procedures
suitable for conditionals in markups.