Let get_bound default to host's bound for sticky spanners
This solves all problems related to the host's right bound being unset
at the point where the Spanner_tracking_engraver tries to pass it on to
the attached sticky grob.  Since bounds can be set as late as in
finalize, the Spanner_tracking_engraver is repurposed to only announce
end, and the bounds are now handled by defaulting to the hosts' when they
are retrieved.

Closes #6158, #6159