Documentation: use paper size a7 for \book example
#6179 observes that we stopped cropping pages in lilypond-book
output. Instead, for this example, the page is really cropped to A8
paper, which is too small for the title "A scale in LilyPond".

Specify A7 paper, which is a large enough for the title. For page
example, it would be best to display a frame around the page, but that
is left for another commit.