Support tail \alternative inside \repeat { }
These now yield the same result:

    \repeat volta \n { \foo \alternative { ... } }
    \repeat volta \n { \foo } \alternative { ... }

Repeat_styler announces the boundaries of repeated sections on behalf of
the iterators of \repeat and \alternative music.

Each instance of Volta_repeat_iterator creates an instance of
Repeat_styler, and each instance of Alternative_sequence_iterator uses
the styler of its enclosing repeat.