DocBook: support tag attributes with a single quote
Before this patch, DocBook files had to use double quotes for
attributes:

	<programlisting language="lilypond">

However, XML makes no distinction with single quotes, like this:

	<programlisting language='lilypond'>

This patch fixes the regexes to allow for both kinds of delimitation of
attributes. The changes onto the DocBook document still use double
quotes, irrespective of what the original document used (as they are the
most common).

Fixes #6185