DocBook: output the raw LilyPond code at the right place
Before this commit, lilypond-book leaves the raw LilyPond book in the XML
file after its processing. If this is the DocBook input:

	<?xml version="1.0" encoding="UTF-8"?>
	<article xml:lang="en_US" xmlns="http://docbook.org/ns/docbook" version="5.2">
	<title>LilyPond-book and LyX</title>
	<mediaobject>
	<textobject>
	<programlisting language="lilypond" role="fragment verbatim staffsize=16
	ragged-right relative=2">
	\relative c'' {  g a b c}
	</programlisting>
	</textobject>
	</mediaobject>
	</article>

The problem, before this commit, is that the <programlisting> tag is
now outside the <mediaobject> and has no particular markup to
eliminate it afterwards (it is forced to be output as a snippet of code,
which is then also printed in the HTML and PDF outputs). Also, its
attributes have been removed in the process (including the LilyPond
language).

With this patch, the <programlisting> is output at the end of the
<mediaobject> tag. It fits the semantics of <mediaobject> very well:
all its direct children are supposed to be alternatives one of the
other (and the raw LilyPond code is definitely the same as an image,
as far as semantics are concerned).

The images _must_ come before the <programlisting> in the XML file,
because DocBook processors are supposed to first display the first
image if possible, the others are only fallbacks.

Fixes #6186