Allow trailing optional arguments to be omitted when called via Scheme
Music functions like `\key` have trailing optional arguments that,
when called via LilyPond input, need to be explicitly omitted using
`\default` to be replaced by their default values.  While in calls
from Scheme, `*unspecified*` can be used equivalently to `\default` in
LilyPond syntax, the end of the argument list in Scheme is explicitly
specified anyway, so `*unspecified*` at this location is redundant.
This change allows it to be omitted in that case.