Vertical_align_engraver: acknowledge hara-kiri-group-spanner-interface
This is a simpler way to detect VerticalAxisGroup than checking for
various conditions on axis groups -- these included a check on the
no-alignment property, which is hereby obviated.  This method is also
less surprising (I remember it took me hours to figure out
no-alignment was the needed ingredient to make
CenteredBarNumberLineSpanner work) and less error-prone (because a
change in the point of time at which an axis group gets its parents
set no longer has a chance of influencing whether that axis group is
added to a VerticalAlignment).