Introduce -dtall-page-formats option
This option sets the format for the "tall page" image, which is used
for online documents. lilypond-book sets it to PNG automatically for
docbook, HTML and texi files. If not specified, the tall page output
is skipped (as itg is not appropriate for PDF-only formats such as LaTeX)

This avoids generating PNG images for each image separately, which is
a side effect of removing framework-eps