Don't give DurationLine a cross-staff callback
It only ever looks at the height of the left item, so cross-staff is
irrelevant. For safety, add a test.

Anecdotally, not marking cross-staff allows the height of the
DurationLine to matter in the distancing between staves, which can
make a difference for an incredibly thick line (visible after \tweak
thickness 100 or so).