Don't check musicality of bounds in line spanner cross-staff determination
This code is primarily for glissandi, but could be relevant for
user-defined line spanners that are not horizontal.  If such a spanner
is cross-staff and broken, it will want to look at the virtual
vertical distance between the left and right note head in order to
determine its slope.  This necessitates marking it cross-staff.  The
code has worked so far because cross-staff was accessed before line
breaking on the original spanner (for VerticalAxisGroup pure heights)
where the bounds could not be non-musical yet, and cached, thus
inherited by the broken pieces.  Make it work by design rather than by
accident.