Handle cross-staff broken glissandi acceptably
The previous code had two paths separated for unclear reasons
depending on the value of the simple-Y property (it seems to have made
sense historically and evolved in such a way that now it doesn't).
One was exercised by glissandi and broke on cross-staff glissandi due
to not communicating refpoints correctly and not having logic for
bringing coordinates on two systems to a common reference.  The other
was exercised by VoiceFollower and FingerGlideSpanner, and could not
handle breaks across more than two systems.  Unify them, fixing the
defects on both sides.

Closes #4806