feta-clefs.mf (new_bulb): Avoid numerical instability
This commit makes the `before` and `after` subpaths longer in a trivial way;
doing so avoids a miscomputation of METAFONT while it tries to find an
intersection between the two paths.  This can be seen in the proof DVI file
of font 'feta11', glyph 'tenor G clef (reduced)', where the bulbs are
deformed otherwise.

Interestingly, METAPOST output is not affected; the final OTF file
`emmentaler-11.otf` doesn't show a distortion for this glyph even without
this patch.