Fix Spanner::find_broken_piece ()
It was finding the first broken piece not less than the specified
system with std::lower_bound, then checking that condition on the
piece, whereas the intended check was evidently that the piece's
system was actually the one requested ('equal' instead of 'not less').
Consequently, it could return a wrong value when no broken piece with
the right system was found.

Fix it by rewriting it as access to the right index in the
broken_intos_ vector, which should also be cheaper than searching
element by element in the vector.