feta-sharps.mf: Tighten bounding boxes vertically
For glyphs with arrows, don't use approximations but base glyph height and
depth on actual values.

The computations now use 'sharpened' values (i.e., variables with a trailing
`#` character to indicate they are not scaled to the output device) for the
bounding boxes.  This has been done for all sharp glyphs; as a side effect,
low-resolution renderings with METAFONT are now much improved and more
consistent.