mf: Restore horizontal metrics of accidentals and/or make them uniform
In the series of commits starting with 28eee886e I was fooled by fake
bounding boxes; my attempts to 'rectify' them by correctly separating
device-independent from device-dependent variables were thus flawed.

This commit restores the width of all accidentals to the values previous to
commit 28eee886e.  Additionally, the breapth values are made uniform, that
is, no increase for arrowed or slashed glyphs (with the exception of the
'3/4 flat' glyph).

A follow-up commit will eventually extend the bounding boxes horizontally,
as originally intended.