mf: Extend bounding boxes horizontally for slashed and arrowed accidentals
This commit makes arrows of accidentals stick out to the left and/or right
of the bounding box in a uniform way, using the arrowed sharp glyphs as a
template.

Use the same offset values for `flat (slashed)` and `flat (slashed twice)`.

Slightly reduce the breapth of the `3/4 flat` to better harmonize with other
flats.

Note that arrowed accidentals of the 'Bravura' font are constructed in a
similar way.