scm/editor.scm: Avoid character ranges in regex
As "man 7 regex" puts it, "Ranges are very collating-sequence-
dependent, and portable programs should avoid relying on them."

This can be noticed starting with Guile 2.0 where make-regexp starts
complaining about "Invalid range end" because UTF-8 locales collate
the DEL character (0x7f) next to the other non-printable characters,
and before '{', so the range {-\x7f is ill-formed.

Fix this issue by taking care of a subset of (printable) characters
that are known to cause problems. Bonus: This includes '`' which was
previously missed (probably by accident).