Don't set an empty track name in MIDI
At least old Solton/Ketron arrangers cannot deal with an explicit
track title of "" in the control track and get mixed up enough that
they refuse playing the whole MIDI file, making "" as the default
track title a liability for users of those instruments.

Strictly speaking, it would be more correct to distinguish the cases
of setting "" as track title explicitly (and pass it on) and not
setting it at all (and omit a track name event) but it is not clear
that it would be of actual value for anybody.