Take slur skylines from stencil
The code to build slur skylines predates the ability to build skylines
from stencils directly. Use this ability, which removes code
duplication. As a bonus, this means that the slur's thickness will
factor into its skylines. This makes for plenty of good-natured
differences in regression tests, where a slight bit of padding is
added in vertical spacing that was previously too tight.