\segnoMark command
The \segnoMark command is like the \mark command, but creates a segno as
a SegnoMark grob.  The argument is a sequence number or \default.

At the start of the score, \segnoMark \default is ignored, but the user
can specify a sequence number to force a segno to appear.

The segnoStyle context property controls whether the command creates
SegnoMark grobs placed outside the staff, or in-staff segno bar lines.

    segnoStyle = #'bar-line
    segnoStyle = #'mark

The \markup \segno command creates a segno scaled for use in a line of
text.