lilypond-book: Do not fail for some `\input` cmds
For includes specified with `\input` that are not found directly,
lilypond-book tries to determine the full path with the `kpsewhich`
utility. If this succeeds, the captured stdout of `kpsewhich` has a trailing
newline appended, which has to be removed before opening the file.

Add a regtest and modify the Makefile in 'input/regression/lilypond-book' to
prepend this directory to 'TEXINPUTS'. This allows testing the `kpsewhich`
lookup mechanism.

Closes #6213.