Doc: Add syntax highlighting
In the HTML documentation, snippets created with `@lilypond[verbatim]`
are now syntax-highlighted using the Pygments library.  This is
implemented through an internal option in lilypond-book.  A CSS
stylesheet is generated by Pygments and the output uses <span> tags
with classes.  Highlighting in PDF output, possibly in
black-and-white, has been deferred.  Info does not support colors.

Because Pygments' support for LilyPond is recent, the library is
vendored in a ZIP archive.  The intent is to keep it so until the last
release has percolated in GNU/Linux distributions we care about.

The font size of code samples, which was somewhat small previously, is
increased slightly to accomodate for the contrast reduction.  The
colors have also been chosen to have good contrast.

At the bottom of each page, there is a clickable link to disable
highlighting.  The setting can be saved permanently using a second
link.

Part of the JavaScript code to turn off syntax highlighting was
authored by Aaron Hill <lilypond@hillvisions.com>.

Closes #1005