Fix dependency on callback order with tablature slurs
The stencil routine was modifying the control points via
`ly:grob-set-property!`.  However, the vertical skylines are not taken
from the stencil, but computed from the control points.  This led to
them being off because they would be constructed from the unmodified
control points.  Even though the next commit will change slurs'
skylines to be based on the stencil, this changes the method to
manipulate the control points directly in that callback.  This should
avoid dependency on callback order in general.