\codaMark command
The \codaMark command is like the \mark command, but creates a coda sign
as a CodaMark grob.  The argument is a sequence number or \default.

At the start of the score, \codaMark \default is ignored, but the user
can specify a sequence number to force the mark to appear.

Two formatting functions are provided:

    codaMarkFormatter = #format-coda-mark     (default)
    codaMarkFormatter = #format-varcoda-mark

The \markup \coda and \markup \varcoda commands create coda signs scaled
for use in a line of text.