\repeat segno
\repeat segno is a folded repeat that notates repetition with a segno
and D.S. or D.C. instructions.  When the following conditions are met,
it also notates alternatives with coda marks in lieu of volta brackets.

* the alternatives appear at the tail of the repeated section
* the alternatives are performed in order (e.g. not 1,3 then 2,4)
* the common part of the repeated section is not empty

Alternative bar numbering is not applied to alternatives notated with
coda marks.  It is applied to alternatives notated with volta brackets
whether they are created by \repeat segno or \repeat volta.

\repeat segno can appear inside \repeat volta, and vice versa; however,
it is ill-advised to combine them such that both create volta brackets,
even if LilyPond engraves what was intended, which it might not.

The context property dalSegnoTextFormatter is a function that creates
a JumpScript for various cases involving D.C., D.S., Fine, Coda, and
other labels.

OUT OF SCOPE / WISH LIST

* use repeat structure, \fine, and \section to improve line breaking
* if the coda section does not begin at a line break, insert an in-staff
  coda mark as in Behind Bars p.239
* add a command to specify jump instructions manually

KNOWN ISSUES / DESIGN CHOICES

* Nesting \repeat segno inside \repeat segno is not supported.
* LilyPond does not infer that \fine applies to the final volta.
  Specify \volta N \fine where necessary to limit its scope.
* LilyPond does not create a thin double bar line at a measure-aligned
  D.S. at the end of the written music.  Specify \section when desired.

Closes #3752