Fix default path insertion for TEXINPUTS et al.
Kpathsea looks for a leading, trailing or double colon in 'TEXINPUTS'
and inserts the next-highest-priority search path there.
It seems the double colon was intended to make sure that
the default search path is always part of 'TEXINPUTS', even if
another path (without trailing colon) is appended to it later.

However, as we always(tm) want to _prepend_ directories to the default
search path, the double colon is not needed.