[DOC] Improve NR simultaneous.itely
Before only Voice is mentioned thoughout the file. This is
misleading up to plain wrong.

The << {…} \\ {…} >> construct creates Bottom contexts of current
type, not only Voices.
In TabStaff we need to use TabVoice or (in case the same music is
used in Staff and TabStaff) we need to use the general
Bottom-context.

Current patch extends explanations and examples to reflect this.