Support overriding dots-direction in \note markup command
In music, dots are shifted (up or down) to avoid them to be centered on a
staff line. In certain situations (for example, dotted notes as denominators
in non-standard time signatures), this is desirable in a \note markup
as well.

Also add regression test input/regression/markup-note-dots-direction.ly
and improve LilyPond coding style in input/regression/markup-note.ly
while at it (use #UP and #DOWN instead of #1 and #-1).