Fix stencil-align-dir-Y of line spanners when Y is nonzero
Only do the translation corresponding to the attachment point of the
line after aligning the stencil on itself.  This fixes
stencil-align-dir-y ironing out Y by re-aligning everything.  In
particular, stencil-align-dir-y is now useful on non-horizontal line
spanners, such as glissandi.  Before this change, the added test would
confusingly print "gliss." at the middle of the staff.