Support C++11 raw string for function and grob interface docstrings
Trim whitespace from the beginning and end of these docstrings in the
autodocumentation infrastructure so that they can be formatted
appropriately.

This was already done for translator descriptions since commit
da88dec516; however, for consistency and to avoid a tiny runtime
overhead, this commit moves it to the documentation autogeneration.