Make smob counting thread-safe
In Guile 2.2, finalizations are executed on a separate thread. This
means that the main rendering thread and the finalizer thread are
racing for access to Smob_core::count.

The impact of losing a few updates is likely small, but for
correctness' sake, use atomic operations to update the counts.

(cherry picked from commit fcc998d849f611ae1c74b661e790a67498abc351)