System: only mark Grob_array if all_elements_scm_ is live
We interface with Guile through Smob objects, which are represented as
(smob_tag, smob_pointer) Scheme cells. The objects are cleaned up
using BDW's finalization feature.

When finalizable objects point to each other, the BDW garbage
collector makes an effort to avoid collecting (and therefore, reusing)
the memory of pointee (this is the "Java finalization" feature). It
uses the GC marking functions for this.

However, the finalizers themselves are not ordered; if there are
dependencies between the objects, it is up to the implementers of the
finalizers to ensure that finalization is properly ordered.

For more background, see the conversation on
https://github.com/ivmai/bdwgc/issues/403

The Guile runtime makes no special effort to order finalizers.

For us, this has two consequences:

* Dead objects may still have their mark functions called.

* Any dependency of a dead object is a valid Guile cell, but if it is a
  smob, the underlying C++ object may already have been deleted. Guile marks
  this with a finalized_smob_tc16 tag.

The current code in System::derived_mark tries to access the C++
object directly, which generates crashes if the System is dead, and
its all_elements_ Grob_array was already finalized.

Many other marking functions access their C++ dependencies directly,
like so

    Book::mark_smob() {
      if (paper_) scm_gc_mark(paper_->self_scm())
      ..
    }

If the dependency was deleted already, self_scm() is a bogus value.
Although this looks questionable (it's essentially a Use-After-Free),
in Guile 2.2, scm_gc_mark() calls BDWGC's GC_mark_and_push() which
sanity checks its input; invalid addresses are thus ignored.

Fixes #6218.

(cherry picked from commit 6555b3841a8d872cb228cbf7247efa34a7bba171)