Add support for Persian music notation
The new file `persian.ly` is based on Kees van den Doel's file with the same
name; while note names are identical, the selection of key and tuning is
different, similar to the input methods used for Turkish and Arabic music.

The documentation of Doel's `persian.ly` has been revised, extended, and
moved to `world.itely`.