Fix footnotes and balloons on MultiMeasureRestNumber and PercentRepeatCounter
The code was taking the offset relative to the left bound of the
Footnote or BalloonText spanner, which is not always the same as its X
parent (for sticky grobs, the latter is always the same as the Y
parent).  This caused wrong placement of balloons on
MultiMeasureRestNumber and PercentRepeatCounter because the offsets of
these spanners relative to their left bounds are nonzero.

Closes #6227