Omit "Fine" text at the written end of the music
Avoid printing "Fine" at the written end of the music, but still print
the final bar line.  The text can be forced to appear by setting the
finalFineTextVisibility context property to #t.

These cases are noteworthy:

* Repeats have been unfolded.  No other repeat notation remains, so
  leaving "Fine" would look strange.

* It is more convenient to code an optionally unfoldable piece as
      \repeat volta 2 { ... } \fine
  than
      \repeat volta 2 { ... \volta 2 \unfolded \bar "|." }

These changes also allow printing both a "D.S."-type instruction and a
"Fine" simultaneously.  The clarity of the score in such a case is
doubtful, but allowing it is now simpler than diagnosing it.