Create all sticky grobs from host's origin engraver
Sticky grobs are particular because their function is closely tied to
their host's.  This creates all of them from the host's origin
engraver, bringing BalloonText and Footnote in consistency with
Parentheses, ControlPoint and ControlPolygon.  This makes it possible
to address overrides for BalloonText and Footnote to the same context
as the one where properties of the annotated grobs are overridden.  It
also fixes vertical spacing of balloons when the Balloon_engraver is
in Score.  Since BalloonText grobs were previously created in Score
even for grobs originating in Staff, they would not be added to the
Staff's VerticalAxisGroup, on contrary to the case where the
Balloon_engraver is consisted in Staff or Voice.

Because footnotes can now be part of individual staves' vertical axis
groups, they need a cross-staff callback, as balloons already had.