Give Footnote, ControlPoint and ControlPolygon cross-staff callbacks
In commit 7aabf0dedf, I thought it was unneeded because they had
extents set to #f.  It turns out that the backend infrastructure
actually doesn't refrain from taking offsets to refpoints even when
the extent is empty.  Footnote works because the grobs are created in
Score, so they are not part of staves' vertical axis groups.  This
will change in the next commit.  ControlPoint and ControlPolygon work
because they attach to grobs without offsets that are problematic in a
cross-staff context.  For safety against implementation changes, give
cross-staff callbacks to all of these.