Fix handling of *unspecified* returned from grob callbacks
This was removing the property entry from the alist completely, which
caused a callback returning *unspecified* to be called on every access
to the property.  This removes the special-casing altogether: this
simplifies the code; the programming error detection was heuristic
anyway; and it primarily stemmed from earlier design rather than a
desire to protect against user errors.

Closes #6226