vowel-transition.cc: Avoid relying on buggy behavior with *unspecified*
Use get_property_data so that callbacks are not triggered.  If they
are, they will in theory not be run again.  This has been working so
far because *unspecified* badly interacted with caching, which is
fixed in the next commit.

Also do a similar fix in lyric-hyphen.cc.  It was not visible due to
additional logic in the print routine (a FIXME is added), but
preventing after-line-breaking from being called before line breaking
still seems like the right thing to do.