Remove Paper_column::before_line_breaking.
This code was introduced in commit 4ed25cec. It apparently tries to
filter out killed Spanners from Paper_column::bounded-by-me. This
might make columns that bound killed spanners not 'used'.

The commit added no regtest that describes when this applies, and we
have no other regtests that are impacted by removing the code.