Volta bracket fixes and enhancements
1. Walk back a misstep from the initial implementation of \volta: the
decision that \volta should always produce a bracket.  Now only
\alternative elements produce brackets.

2. Enhance repeat-related iterators to offer some support for nesting
\alternative within \alternative.  This gives back some of what was
taken from \volta.

3. Enhance Volta_engraver to handle nested repeats.  It is not currently
a goal to engrave well when multiple brackets are live.  The goal is to
avoid combining info from different layers of the repeat structure in a
way that might cause a VoltaBracket grob to be created at the wrong time
or to fail to be created.

4. Update Timing_translator to base alternative bar numbers on the
outermost bracketed alternative only.

Ideas for further clean-up:
* Volta_engraver code to handle overlapping volta spans at a given depth
  is probably unnecessary now.  When touching this logic, it would be
  good to try automating some of the manual stuff in volta-overhang.ly.
* We probably don't need both AlternativeEvent and VoltaSpanEvent, and
  we probably don't even need Volta_specced_music_iterator.