Fix broken spanner functions
They were broken (pun intended) on spanners of which one bound is a
non-musical column at a system boundary.  For example,
unbroken-or-last-broken-spanner? would return false on a spanner
running to the NonMusicalPaperColumn at the end of a score.  This
fixes the known issue with duration lines running to the end of the
piece, and will be needed in a refactoring of spanner-placement
treatment for balloon-interface.

While at it, document these functions.