Fix BalloonText width & height calculations
Balloons were not working on stems and stem tremolos because the
heights of these items depend on line breaking.  This means that they
need pure heights, even though they are items.  They also need a
width, which cannot be stencil-based, but can be computed through a
skewed stencil like is done for pure heights because they only depend
on the host's width, not on its height.

Closes #6155