Remove System members pointing to all-elements
The variable all_elements_scm_ is not involved in hot code paths, so
avoid the cognitive overhead of specially marking the all-elements
Grob_array.

Use some C++11 for loops for improved readability.

This results in some extra property access, but benchmarking has shown
that the impact on run time is neglible.