Disentangle TupletBracket creation
Previously, Tuplet_bracket::parallel_beam used to do the check for
equal length of Beam and TupletBracket as a side effect. This necessitated
the creation of dummy booleans in Tuplet_bracket::print and
Tuplet_bracket::calc_position_and_height.

This commit moves the length comparison to a new function
Beam::is_equally_long in preparation for a future addition of additional
options for TupletBracket.bracket-visibility (the allowed values of which
are currently a boolean? or 'if-no-beam).