Fix wrong use of `scm_is_rational`
Guile's notion of "rational" includes floating point numbers ("inexact
rationals") other than ±NaN and ±inf.  That doesn't work for LilyPond
which has `Rational` as an _exact_ rational but _including_ ±inf while
excluding ±NaN.  Any use of `scm_is_rational` is likely to be a
mistake in the LilyPond code base (the code for the predicate
`is_scm<Rational>` could be an exception but happens to be implemented
without using `scm_is_rational` as well).