Doc-IR: fix references to class-specific grob interfaces
This corrects an oversight from commit 83faf57ecc.  It generated, for
example, `This object is of class @iref{spanner-interface,Spanner}.`,
with the intent being to point at the relevant page for learning about
spanners.  I overlooked that `@iref` is a macro doing special things
in TeX for PDF output, and does not support the two-argument form.
This led to PDF output like "This object is of class
spanner-interface,Spanner (page ⟨undeﬁned⟩)."  Instead, stay with one
argument and print "This object is of class Spanner (characterized by
spanner-interface)".  For sticky grobs, this looks like: "This object
can be of either of the following classes: Item (characterized by
item-interface) or Spanner (characterized by spanner-interface). It
supports the following interfaces conditionally depending on the
class: item-interface (page 575), and spanner-interface (page 607)."