Add markup->string tests
In addition to being public, this function is used for PDF metadata
and MIDI text such as lyric and markers.  Metadata are not tested at
all by the current infrastructure.  MIDI output is to some extent, but
the results are not shown in regression test diffs (only in a log).
Thus, it seems better to add testing coverage for markup->string
itself.