Rework markup->string
markup->string had a number of problems.

- First and foremost, it processed all bindings in the (lily) module
  at every invocation.  This led to poor speed of MIDI output for
  scores with lyrics because markup->string is called on every lyric
  word since commit a89e32564f.

- There were also oddities in the code: for example, precomputing the
  list of markup commands having at least one markup?  or markup-list?
  argument seems to have been unnecessary because it was just taking
  the last argument and converting it, returning "" if that argument
  was not a markup.  Also, the result of last-pair is always 0 and 1,
  so the code path checking for a length greater than 1 was dead.

- It did not work for markup arguments that were not last.  For
  example, (markup->string #{ \markup \combine A B #}) returned "B".

- Finally, the special cases were maintained inside the function
  itself, and it could not work for custom markup commands at all.

This completely reimplements markup->string by making it look up a
"to-string" handler on the markup command.  define-markup-command
allows specifying it by keyword argument, like this:

```
(define-markup-command (foo layout props arg1 arg2 ...)
                       (type1? type2? ...)
  #:as-string ...
  ...)
```

markup->string evaluates the #:as-string argument in the context of
the same bindings that are available in the body of the markup
command, namely layout, props, the arguments, and optionally the
properties.  If not specified, the default is to join the results of
markup->string on all markup or markup list arguments by spaces.
(begin ...)  if needed.

The optional arguments to markup->string are now layout and props as
for markup interpretation.  layout defaults to #f and props to '().
Thus, if the handler uses the layout, it should check for it being
non-#f.

A number of handlers are added, mostly to replace the existing special
cases or to cater for the default now being to consider all arguments
and not just the last one.

A convert-ly rule is added to convert existing uses of markup->string
(it does not bother limiting this conversion to uses of the
two-argument form).

Closes #6257