docker: Add base image
This image contains all packages necessary to build LilyPond, but
does not drop privileges to nobody for GitLab CI. It can and will
be used as a starting point for other images, for example to build
the documentation of releases.

Also rebuild the ci/ubuntu-18.04.guile2 image based on it.