doc: Prevent gs from converting PNG to JPG
If `extractpdfmark` is used, an additional conversion step using GhostScript
is necessary.  By default, `gs` performs downsampling of PNG images, which
are then compressed with the JPEG algorithm.  Especially while viewing the
created PDF documents online this yields ugly artifacts for screenshots and
the like.

This commit adds a bunch of command line flags to `gs` that prevent
downsampling and JPEG encoding.