output-distance: extend size of input images
When using to ImageMagick for diffing images, we must have the union
of both image sizes. When taking the intersection, cropping may cut
off dissimilar regions.

We must diff on equally sized images, as older versions of ImageMagick
(e.g. the one on Ubuntu 18.04) cannot compare images of differing sizes.

Skip the resizing if the image has correct size. This is a significant
time saver for diffing large sets of images.