Limit bar line extent to 2 staff-spaces by default
The coping mechanism for choosing the height of bar lines on vertically
short staves yields 2-space-high bar lines on a 1-line staff.  Applying
it only to staves shorter than 1 space created the strange situation
that increasing the number of staff lines to 2 reduced the height of the
bar lines, harming readability.

This change extends bar lines if they are less than 2 staff-spaces.

Testing exposed the separate issue that tick bar lines positioned
themselves relative to the top of normal bar lines rather than the
top line of the staff.  (The outcome is often the same.)  This change
also fixes that.

Closes #6266.