Reimplement \bar with an event
The \bar command used to be implemented as \set Timing.whichBar = ...,
but now it sends an event and Timing_translator takes care of setting
both the whichBar property and a new barForced property.  The barForced
property allows engravers to tell that whichBar was set by the \bar
command rather than by another internal source.  Setting whichBar
directly should still work as it did before, but it is now documented as
deprecated.

Commit 59e2179bab3ad64922e1e6e22dbbb636afb7975c improved a spacing issue
related to default bar lines in ancient music, with the slight drawback
that a user couldn't get an effect from an explicit \bar "|" even if it
was desired.  Now, Bar_engraver consults barForced before adjusting the
bar line to suit its local context; it uses a forced bar line as-is.

The possibilities for using commands like \repeat and \section in
ancient music are expanded by mapping more bar types in Bar_engraver. As
noted in comments, we will probably want a better solution before long,
but this is an outward improvement that allows us to add some tests.